Fix minor points introduced by pull #753 for long indexing

+++ nd4j-backends/nd4j-api-parent/nd4j-api/src/main/java/org/nd4j/linalg/api/ops/impl/accum/distances/CosineSimilarity.java
@@ -188,7 +188,7 @@ public class CosineSimilarity extends BaseAccumulation {
     @Override
     public void exec(int... dimension){
         int[] retShape = ArrayUtil.removeIndex(x.shape(), dimension);
-        long nOps = x.tensorssAlongDimension(dimension);
+        int nOps = x.tensorssAlongDimension(dimension);
         z = Nd4j.create(retShape);
         for( int i = 0; i < nOps; i++ ){
             double d = Nd4j.getExecutioner().execAndReturn((CosineSimilarity) opForDimension(i,dimension)).getFinalResult().doubleValue();
